# go-commons/gen/openapiv2/proto/route/route.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/route/route.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Route"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/route/end": {
      "post": {
        "operationId": "Route_EndSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/locationDeviceID"
            }
          }
        ],
        "tags": [
          "Route"
        ]
      }
    },
    "/v1/route/location": {
      "post": {
        "operationId": "Route_RouteChat",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/locationOBUData"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of locationOBUData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/locationOBUData"
            }
          }
        ],
        "tags": [
          "Route"
        ]
      }
    },
    "/v1/route/start": {
      "post": {
        "operationId": "Route_StartSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/locationInitResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/locationDeviceID"
            }
          }
        ],
        "tags": [
          "Route"
        ]
      }
    }
  },
  "definitions": {
    "locationDeviceID": {
      "type": "object",
      "properties": {
        "deviceId": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "locationInitResponse": {
      "type": "object",
      "properties": {
        "sessionId": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "locationOBUData": {
      "type": "object",
      "properties": {
        "deviceId": {
          "type": "string",
          "format": "byte",
          "title": "16-byte uuid"
        },
        "latitude": {
          "type": "number",
          "format": "double"
        },
        "longitude": {
          "type": "number",
          "format": "double"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}